Fixed Ref_Seq drop-out
Well's aligning to the first entry in input.fasta would not join in the
reference sequence from input-refs.tsv. Forgot to specify an implicit
header in mlr, which would set the first line to the header. If the user
only specified one reference in input.fasta, all ref_seq rows in
aggregated-stats.tsv would get set to NA.